Fix DeepScan issues
