+++ packages/views/screens/managers/dashboard/infoBanner/infoBanner.js
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from 'react';
 import { withTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 import Icon from 'src/theme/Icon';
-import Illustration from 'src/modules/common/components/illustration';
 import styles from './infoBanner.css';
 
 const InfoBanner = ({
