+++ packages/views/screens/managers/request/request.test.js
@@ -1,8 +1,7 @@
 import React from 'react';
-import { tokenMap } from '@token/fungible/consts/tokens';
 import { mountWithRouter } from 'src/utils/testHelpers';
 import accounts from '@tests/constants/wallets';
-import Request from '.';
+import Request from './request';
 
 jest.mock('@shared/converter', () => (
   function ConverterMock() {
@@ -14,98 +13,29 @@ describe('Request', () => {
   let wrapper;
 
   const props = {
-    account: {
-      info: {
-        LSK: {
-          summary: { address: accounts.genesis.summary.address },
-        },
-      },
-    },
-    token: tokenMap.LSK.key,
+    address: accounts.genesis.summary.address,
     t: v => v,
-    history: {
-      push: jest.fn(),
-      location: {
-        pathname: 'wallet',
-        search: '?modal=request',
-      },
-    },
+  };
+  const routeConfig = {
+    pathname: 'wallet',
+    search: '?modal=request',
   };
 
   beforeEach(() => {
-    wrapper = mountWithRouter(Request, props, props.history.location);
+    wrapper = mountWithRouter(Request, props, routeConfig);
   });
 
-  describe('Amount field', () => {
-    it('Should show converter', () => {
-      expect(wrapper.find('.amount .converted-price')).toExist();
-    });
+  it('Should update share link with amount and reference', () => {
+    const shareLink = `lisk://wallet/send?recipient=${props.address}`;
+    let evt;
+    expect(wrapper.find(Request).state('shareLink')).toMatch(shareLink);
 
-    it('Should add leading 0 if . is inserted as first character', () => {
-      const evt = { target: { name: 'amount', value: '.1' } };
-      wrapper.find('.fieldGroup input').simulate('change', evt);
-      wrapper.update();
-      expect(wrapper.find('.fieldGroup input').props().value).toEqual('0.1');
-    });
+    evt = { target: { name: 'reference', value: 'test' } };
+    wrapper.find('.fieldGroup').at(1).find('AutoResizeTextarea').simulate('change', evt);
+    expect(wrapper.find(Request).state('shareLink')).toMatch(`${shareLink}&${evt.target.name}=${evt.target.value}`);
 
-    it('Should show error feedback if letters inserted', () => {
-      const evt = { target: { name: 'amount', value: 'abc' } };
-      expect(wrapper.find('.amount Feedback')).toHaveText('');
-      wrapper.find('.fieldGroup input').simulate('change', evt);
-      wrapper.update();
-      expect(wrapper.find('.amount Feedback')).toHaveText('Provide a correct amount of LSK');
-    });
-
-    it('Should show error feedback if ending in . or multiples .', () => {
-      const multipleDotsEvt = { target: { name: 'amount', value: '1.2.3' } };
-      const amountField = wrapper.find('.fieldGroup').at(0);
-      wrapper.update();
-      expect(wrapper.find('.amount Feedback')).toHaveText('');
-      amountField.find('input').simulate('change', multipleDotsEvt);
-      wrapper.update();
-      expect(wrapper.find('.amount Feedback')).toHaveText('Provide a correct amount of LSK');
-    });
-  });
-
-  describe('Reference field', () => {
-    it('Should show feedback if some text inserted and hide if empty', () => {
-      const referenceField = wrapper.find('.fieldGroup').at(1);
-      let evt = { target: { name: 'reference', value: 'test' } };
-      expect(wrapper.find('.fieldGroup .feedback.show')).not.toExist();
-      referenceField.find('AutoResizeTextarea').simulate('change', evt);
-      wrapper.update();
-      expect(wrapper.find('.fieldGroup .feedback')).toExist();
-      expect(wrapper.find('.fieldGroup .feedback.error')).not.toExist();
-
-      evt = { target: { name: 'reference', value: '' } };
-      referenceField.find('AutoResizeTextarea').simulate('change', evt);
-      wrapper.update();
-      expect(wrapper.find('.fieldGroup .feedback.show')).not.toExist();
-    });
-
-    it('Should show error feedback over limit of characters', () => {
-      const referenceField = wrapper.find('.fieldGroup').at(1);
-      const evt = {
-        target: {
-          name: 'reference',
-          value: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit volutpat.',
-        },
-      };
-      referenceField.find('AutoResizeTextarea').simulate('change', evt);
-      wrapper.update();
-      expect(wrapper.find('.fieldGroup .feedback')).toExist();
-      expect(wrapper.find('.fieldGroup .feedback.error')).toExist();
-    });
-  });
-
-  describe('Share Link', () => {
-    it('Should copy and set timeout on click', () => {
-      expect(wrapper.find('.copy-button button')).not.toBeDisabled();
-      wrapper.find('.copy-button button').simulate('click');
-      expect(wrapper.find('.copy-button button')).toBeDisabled();
-      jest.advanceTimersByTime(3100);
-      wrapper.update();
-      expect(wrapper.find('.copy-button button')).not.toBeDisabled();
-    });
+    evt = { target: { name: 'amount', value: 1 } };
+    wrapper.find('.fieldGroup').at(0).find('input').simulate('change', evt);
+    expect(wrapper.find(Request).state('shareLink')).toMatch(`${shareLink}&${evt.target.name}=${evt.target.value}`);
   });
 });
