+++ src/modules/transaction/manager/explorerManager.js
@@ -19,8 +19,8 @@ const defaultSort = 'timestamp:desc';
 export default compose(
   withData({
     transactions: {
-      apiUtil: (network, { token, ...params }) =>
-        getTransactions({ network, params: normalizeTransactionParams(params) }, token),
+      apiUtil: (network, params) =>
+        getTransactions({ network, params: normalizeTransactionParams(params) }),
       getApiParams: (_, { address, sort }) => ({
         address,
         sort,
